requirements.txt
streamlit
pandas
numpy
transformers
torch
plotly
scikit-learn
google-generativeai
